+++ src/test/java/ua/shpp/eqbot/service/UserServiceIntegrationTest.java
@@ -14,7 +14,7 @@ import ua.shpp.eqbot.model.UserEntity;
 import ua.shpp.eqbot.repository.ProviderRepository;
 import ua.shpp.eqbot.repository.UserRepository;
 import ua.shpp.eqbot.telegrambot.EqTelegramBot;
-import ua.shpp.eqbot.utility.UserMapper;
+import ua.shpp.eqbot.mapper.UserMapper;
 import ua.shpp.eqbot.validation.UserValidateService;
 
 import java.time.LocalDateTime;
